+++ src/Fable.Cli/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* [JS/TS] Add `Async.AwaitEvent` support (fixes #4239)
+
 ### Fixed
 
 * [Dart] Enable explicit variable typing for more data types (by @ncave)